@@ -1,7 +1,7 @@
 //= require jquery
 //= require jquery_ujs
-//= require jquery.ui.core
-//= require jquery.ui.tabs
+//= require jquery-ui/core
+//= require jquery-ui/tabs
 //= require galleria
 //= require jquery.clearfield
 //= require jquery.purr
