@@ -1,8 +1,8 @@
 # -*- encoding : utf-8 -*-
 class Flicker < ActiveRecord::Base
   belongs_to :event
-  scope :event, proc { |event| where(:event_id => event) } 
-  attr_accessible :include_in_carousel, :event_id, :image_file_name, :image_file_size, :is_video, :name, :creator, :hostid, :image, :description,  :image_content_type
+  scope :event, -> (e) {  where(:event_id => e) } 
+  # attr_accessible :include_in_carousel, :event_id, :image_file_name, :image_file_size, :is_video, :name, :creator, :hostid, :image, :description,  :image_content_type
   has_attached_file :image, 
                     :path =>  ":rails_root/public/images/contrib/:id/:style/:basename.:extension", 
                     :url => "/images/contrib/:id/:style/:basename.:extension",
