@@ -65,7 +65,7 @@ module Ptarmigan
     # This is necessary if your schema can't be completely dumped by the schema dumper,
     # like if you have constraints or database-specific column types
     # config.active_record.schema_format = :sql
-    config.active_record.whitelist_attributes = false
+    # config.active_record.whitelist_attributes = false
 
 
     # Enable the asset pipeline
