@@ -22,17 +22,18 @@ Ptarmigan::Application.configure do
 
   # Only use best-standards-support built into browsers
   config.action_dispatch.best_standards_support = :builtin
+  config.active_record.raise_in_transactional_callbacks = true
 
   # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
+  config.eager_load = false
 
   # Log the query plan for queries taking more than this (works
   # with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 0.5
+
   config.action_mailer.delivery_method = :letter_opener
   config.action_mailer.default_url_options = { :host => 'localhost:3000' }
   config.assets.compress = false
-  config.serve_static_assets = true
+  config.serve_static_files = true
   # Expands the lines which load the assets
   config.assets.debug = false
   config.action_controller.asset_host = Proc.new { |source|
